feat: evolve crown jewels to more flexible targets (#35)
This implements a more generic form of target to include in operations.
Previously, there was only "Crown Jewels". Now there are "Targets" with
a bool to flag them as Crown Jewels.

There are also numerous additional UI components to highlight where
assets are targeted in operations and specific techniques, and to better
measure the specific outcomes of individual targets.